% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv_genes.R
\name{hv_genes}
\alias{hv_genes}
\title{graph visualization and community detection}
\usage{
hv_genes(df, n = 100)
}
\arguments{
\item{df}{An arbitrary graph object in R.}

\item{n}{Set it TRUE when the graph is directed. (default = FALSE)}
}
\value{
If plot = TRUE it plots the non-interactive graph (If plot.community = TRUE plots communities too) also returns a list contains:
\item{graph}{an igraph object.}
\item{betweenness}{betweenness measurements of each node.}
\item{network}{a visNetwork plot of the graph.}
\item{communities}{a named vector indicating the community of each node.}
}
\description{
Converts the graph to an igraph object, finds communities and plots it using qgraph package.
}
\section{Additional arguments}{

\describe{
\item{groups}{A list that indicates which community each node is. The automatic community detection will be ignored when it is set.}
\item{plot.community}{Logical indicating if communities should be plotted. Defaults to FALSE.}
\item{filename}{Name of the plot file without extension. (qgraph function argument)}
\item{filetype}{A character indicates the file type to save the plots in. (qgraph function argument)}
\item{community.list}{A list indicates which communities should be plotted. When is not set, will plot all the communities.}
}
}

\examples{
require(datasets)
data("Harman23.cor")
gv = graph_vis(Harman23.cor$cov, plot = TRUE, plot.community = TRUE, community.list = c(1, 2))

}
\author{
Elyas Heidari
}
