% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sel.R
\name{sel_genes}
\alias{sel_genes}
\title{Select genes for the network construction}
\usage{
sel_genes(data, method = "hvg", n = 300, spec = "human")
}
\arguments{
\item{data}{the count matrix with gene symbols as rownames.}

\item{method}{a string or list of strings indicate methods used to select genes. See the details for more information. (default = 'hvg')}

\item{n}{number of genes to be selected.}

\item{spec}{the species, used just when 'tf' is used as a method.}
}
\value{
a vector of gene symbols as strings.
}
\description{
Subsets genes to be used as the network nodes
}
\details{
for method, 'hvg': highly variable genes, 'tf': transcription factors, 'rand': random subsetting.
}
\author{
Elyas Heidari
}
